requirements.txt
aiodns==3.1.1
aiohttp==3.9.1
aiohttp-socks==0.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
black @ file:///Users/runner/miniforge3/conda-bld/black-recipe_1701335926439/work
Brotli==1.1.0
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click @ file:///home/conda/feedstock_root/build_artifacts/click_1692311806742/work
colorama==0.4.6
distro==1.8.0
executing==2.0.1
flake8 @ file:///home/conda/feedstock_root/build_artifacts/flake8_1690833614949/work
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
icecream==2.1.3
idna==3.6
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
libgen-api==1.0.1
lxml==4.9.4
mccabe @ file:///home/conda/feedstock_root/build_artifacts/mccabe_1643049622439/work
multidict==6.0.4
mypy-extensions @ file:///home/conda/feedstock_root/build_artifacts/mypy_extensions_1675543315189/work
numpy==1.26.2
openai==1.6.0
outcome==1.3.0.post0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.1.4
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathspec @ file:///home/conda/feedstock_root/build_artifacts/pathspec_1702249949303/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1701708255999/work
pycares==4.4.0
pycodestyle @ file:///home/conda/feedstock_root/build_artifacts/pycodestyle_1697202867721/work
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes @ file:///home/conda/feedstock_root/build_artifacts/pyflakes_1690656279496/work
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv-split_1693865766170/work
python-http-client==3.3.7
python-socks==2.4.4
pytz==2023.3.post1
requests==2.31.0
selenium==4.16.0
sendgrid==6.11.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starkbank-ecdsa==2.2.0
tqdm==4.66.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.9.0
urllib3==2.1.0
wsproto==1.2.0
yarl==1.9.4
